requirements.txt
markdown>=3.3.0
weasyprint>=52.5
PyPDF2>=2.0.0
pyyaml>=6.0
html2text>=2020.1.16
odfpy>=1.4.1
Pillow>=9.0.0